requirements.txt
pytorch-lightning
hydra-core
omegaconf
opencv-python
albumentations
numpy
torchmetrics[image]
wandb
matplotlib